fix: add lookup before creating cloudflare cronjob symlink and comment out inital geoipupdate

+++ src/planet-4-151612/openresty/bin/nginx_configure_cloudflare.sh
@@ -59,4 +59,10 @@ CRON_SCHEDULE="cron.daily"
 CLOUDFLARE_DAILY_CRON_FILE_PATH="/etc/$CRON_SCHEDULE/nginx_update_cloudflare_ips"
 CLOUDFLARE_CRON_FILE="/app/bin/nginx_update_cloudflare_ips.sh"
 
-ln -s $CLOUDFLARE_CRON_FILE $CLOUDFLARE_DAILY_CRON_FILE_PATH
+# Only create the symbolic link if it doesn't already exist
+if [ ! -L "$CLOUDFLARE_DAILY_CRON_FILE_PATH" ]; then
+  ln -s $CLOUDFLARE_CRON_FILE $CLOUDFLARE_DAILY_CRON_FILE_PATH
+  echo "Symlink created at $CLOUDFLARE_DAILY_CRON_FILE_PATH"
+else
+  echo "Symlink already exists at $CLOUDFLARE_DAILY_CRON_FILE_PATH"
+fi
